refactor: updated asset freeze txn to conform to other txns
